Update MediaSingle.php
attrs.width is not an int but float
https://developer.atlassian.com/cloud/jira/platform/apis/document/nodes/mediaSingle/
"attrs.width number Floating point number between 0 and 100"

+++ src/Node/Block/MediaSingle.php
@@ -31,9 +31,9 @@ class MediaSingle extends BlockNode implements JsonSerializable
         Media::class,
     ];
     private string $layout;
-    private ?int $width;
+    private ?float $width;
 
-    public function __construct(string $layout, ?int $width = null, ?BlockNode $parent = null)
+    public function __construct(string $layout, ?float $width = null, ?BlockNode $parent = null)
     {
         if (!\in_array($layout, [
             self::LAYOUT_WRAP_LEFT,
@@ -77,7 +77,7 @@ class MediaSingle extends BlockNode implements JsonSerializable
         return $this->layout;
     }
 
-    public function getWidth(): ?int
+    public function getWidth(): ?float
     {
         return $this->width;
     }
